backend/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
marshmallow==3.20.1
python-dotenv==1.0.0
pytest==7.4.0
gunicorn==21.2.0